Support bulk import with folders.
This feature adds the ability for the "Import" operation to import
folders.  A folder JSON is detected through the presence of the
`name`, `type`, and `sorting` fields.

This is intended for use with `package`-style imports, where the
import JSON looks like:
```
{
  "package": { },
  "items": [
    { <Single Document 1> },
    { <Single Document 2> },
    { ... },
    { <Single Document N> },
  ],
  "folders": [
    { <Folder Document 1> },
    { <Folder Document 2> },
    { ... },
    { <Folder Document N> },
  ]
}
```

The import will now gather both the `items` and `folders` array for
import.

In order for an import to place files into folders, the import needs
to maintain the Foundry IDs.  This means that the 'Keep IDs on Import'
setting is mandatory for folder imports to work correctly.

If this option is not enabled, all the folders will be imported but
assigned new IDs.